Move DbFindingTypeKey to its own file
