+++ feat/findings/be/driven/impl/src/main/kotlin/cz/adamec/timotej/snag/findings/be/driven/impl/internal/DbFindingTypeKey.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright (c) 2026 Timotej Adamec
+ * SPDX-License-Identifier: MIT
+ *
+ * This file is part of the thesis:
+ * "Multiplatform snagging system with code sharing maximisation"
+ *
+ * Czech Technical University in Prague
+ * Faculty of Information Technology
+ * Department of Software Engineering
+ */
+
+package cz.adamec.timotej.snag.findings.be.driven.impl.internal
+
+import cz.adamec.timotej.snag.feat.findings.business.FindingType
+
+internal enum class DbFindingTypeKey {
+    CLASSIC,
+    UNVISITED,
+    NOTE,
+}
+
+internal fun FindingType.toDbKey(): DbFindingTypeKey =
+    when (this) {
+        is FindingType.Classic -> DbFindingTypeKey.CLASSIC
+        is FindingType.Unvisited -> DbFindingTypeKey.UNVISITED
+        is FindingType.Note -> DbFindingTypeKey.NOTE
+    }
